+++ app/Http/Requests/authUser/RegisterUserRequest.php
@@ -1,0 +1,45 @@
+<?php
+
+namespace App\Http\Requests\authUser;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class RegisterUserRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     */
+    public function authorize(): bool
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
+     */
+    public function rules(): array
+    {
+        return [
+            'username' => 'required|string|min:4|max:30',
+            'email' => 'required|email|max:255|unique:users,email',
+            'password' => 'required|min:6|regex:/^(?=.*[0-9])(?=.*[\W_]).+$/|confirmed',
+        ];
+    }
+
+    public function messages(): array
+    {
+        return [
+            'username.required' => 'Username is required',
+            'username.min' => 'Username must be at least 6 characters',
+            'username.max' => 'Username must be less than 30 characters',
+            'email.required' => 'Email is required',
+            'email.email' => 'Invalid email format',
+            'email.max' => 'Email must be less than 30 characters',
+            'email.unique' => 'Email already exists',
+            'password.required' => 'Password is required',
+            'password.min' => 'Password must be at least 6 characters',
+        ];
+    }
+}